+++ Views/IVA/index.php
@@ -4,10 +4,12 @@
     <div class="row" style="height:95.3245%; background-color: #e8e8e9;">
         <div class="row" style="margin-top: 5%;">
             <div class="col-sm-12">
-                <form action="router.php?c=iva&a=pesquisar" method="POST">
-                    <input type="text" name="pesquisa" style="float:left; margin-bottom:2vh; width:15vw" placeholder="Pesquisar IVA" class="form-control bg-dark text-white">
+                <form action="" method="GET">
+                    <input type="text" name="pesquisa" style="float:left; margin-bottom:2vh; width:15vw" value="<?php if (isset($pesquisa)) echo $pesquisa; ?>" placeholder="Pesquisar IVA" class="form-control bg-dark text-white">
+                    <input type="hidden" name="c" value="iva">
+                    <input type="hidden" name="a" value="pesquisar">
                     <button type="submit" style="float:left; margin-bottom:2vh;" class="btn btn-dark"><i class="bi bi-search"></i></button>
-                    <?php if ($pesquisa == true) { ?>
+                    <?php if (isset($pesquisa)) { ?>
                         <a href="router.php?c=iva&a=index" style="float:left; margin-bottom:2vh;">Limpar Pesquisa</a>
                     <?php } ?>
                 </form>
@@ -30,7 +32,7 @@
                     </thead>
                     <tbody>
                         <?php foreach ($ivas as $iva) { ?>
-                            <tr>
+                            <tr <?php if (isset($errors) && $errors == $iva->id) echo ' style="border: 2px solid red"'; ?>>
                                 <td><?= $iva->id ?></td>
                                 <td><?= $iva->descricao ?></td>
                                 <td style="text-align: right; width:10%"><?= $iva->percentagem ?>%</td>
@@ -40,9 +42,10 @@
                                                                     echo 'Não';
                                                                 }  ?></td>
                                 <td style="text-align:right; padding-right:2%">
+                                    <?php if (isset($errors) && $errors == $iva->id) echo '<span class="text-danger">Tem de apagar os produtos associados!</span>' ?>
                                     <a href="router.php?c=iva&a=show&id=<?= $iva->id ?>" class="text-black" title="Mostrar"><i class="fs-4 bi bi-eye"></i></a>
-                                    <a href="router.php?c=iva&a=edit&id=<?= $iva->id ?>" class="text-black" title="Mostrar"><i class="fs-4 bi bi-pencil" title="Editar"></i></a>
-                                    <a href="router.php?c=iva&a=delete&id=<?= $iva->id ?>" class="text-black" title="Apagar"><i class="fs-4 bi bi-trash"></i></a>
+                                    <a href="router.php?c=iva&a=edit&id=<?= $iva->id ?>" class="text-black" title="Editar"><i class="fs-4 bi bi-pencil" title="Editar"></i></a>
+                                    <a onclick="return confirm('Tem a certeza que quer apagar?');" href="router.php?c=iva&a=delete&id=<?= $iva->id ?>" class="text-black" title="Apagar"><i class="fs-4 bi bi-trash"></i></a>
                                 </td>
                             </tr>
                         <?php } ?>
