+++ Views/Produtos/index.php
@@ -4,11 +4,13 @@
     <div class="row" style="height:95.3245%; background-color: #e8e8e9;">
         <div class="row" style="margin-top: 5%;">
             <div class="col-sm-12">
-                <form action="router.php?c=produto&a=pesquisar" method="POST">
-                    <input type="text" name="pesquisa" style="float:left; margin-bottom:2vh; width:15vw" placeholder="Pesquisar produto" class="form-control bg-dark text-white">
+                <form action="" method="GET">
+                    <input type="text" name="pesquisa" style="float:left; margin-bottom:2vh; width:15vw" value="<?php if (isset($pesquisa)) echo $pesquisa; ?>" placeholder="Pesquisar produto" class="form-control bg-dark text-white">
                     <input type="hidden" name="place" value="index">
+                    <input type="hidden" name="c" value="produto">
+                    <input type="hidden" name="a" value="pesquisar">
                     <button type="submit" style="float:left; margin-bottom:2vh;" class="btn btn-dark"><i class="bi bi-search"></i></button>
-                    <?php if ($pesquisa == true) { ?>
+                    <?php if (isset($pesquisa)) { ?>
                         <a href="router.php?c=produto&a=index" style="float:left; margin-bottom:2vh;">Limpar Pesquisa</a>
                     <?php } ?>
                 </form>
@@ -19,19 +21,35 @@
                             <h3>Referencia</h3>
                         </th>
                         <th>
-                            <h3>Descrição</h3>
+                            <a class="text-decoration-none text-white" href="router.php?c=produto&a=order&col=descricao&order=<?php if (!isset($descricao)) echo 'asc';
+                                                                                                                                else if ($descricao == 'asc') echo 'desc';
+                                                                                                                                else echo 'asc' ?>&pesquisa=<?php if (isset($pesquisa)) echo $pesquisa; ?>">
+                                <h3>Descricao <?php if (isset($descricao)) { ?><i class="bi bi-caret-<?php if ($descricao == 'asc') echo 'up';
+                                                                                                        else echo 'down'; ?>-fill fs-5"></i><?php } ?></h3>
+                            </a>
                         </th>
                         <th>
-                            <h3>Preço</h3>
+                            <a class="text-decoration-none text-white" href="router.php?c=produto&a=order&col=preco&order=<?php if (!isset($preco)) echo 'asc';
+                                                                                                                            else if ($preco == 'asc') echo 'desc';
+                                                                                                                            else echo 'asc' ?>&pesquisa=<?php if (isset($pesquisa)) echo $pesquisa; ?>">
+                                <h3>Preço <?php if (isset($preco)) { ?><i class="bi bi-caret-<?php if ($preco == 'asc') echo 'up';
+                                                                                                else echo 'down'; ?>-fill fs-5"></i><?php } ?></h3>
+                            </a>
                         </th>
                         <th>
-                            <h3>Stock</h3>
+                            <a class="text-decoration-none text-white" href="router.php?c=produto&a=order&col=stock&order=<?php if (!isset($stock)) echo 'asc';
+                                                                                                                            else if ($stock == 'asc') echo 'desc';
+                                                                                                                            else echo 'asc' ?>&pesquisa=<?php if (isset($pesquisa)) echo $pesquisa; ?>">
+                                <h3>Stock <?php if (isset($stock)) { ?><i class="bi bi-caret-<?php if ($stock == 'asc') echo 'up';
+                                                                                                else echo 'down'; ?>-fill fs-5"></i><?php } ?></h3>
+                            </a>
                         </th>
                         <th>
                             <h3>IVA</h3>
                         </th>
                         <th>
-                            <h3>IVA %</h3>
+                            <h3>IVA(%)</h3>
+                            </a>
                         </th>
                         <th style="border-radius: 0 10px 0 0;"></th>
                     </thead>
@@ -47,7 +65,7 @@
                                 <td>
                                     <a href="router.php?c=produto&a=show&id=<?= $produto->id ?>" class="text-black" title="Mostrar"><i class="fs-4 bi bi-eye"></i></a>
                                     <a href="router.php?c=produto&a=edit&id=<?= $produto->id ?>" class="text-black" title="Mostrar"><i class="fs-4 bi bi-pencil" title="Editar"></i></a>
-                                    <a href="router.php?c=produto&a=delete&id=<?= $produto->id ?>" class="text-black" title="Apagar"><i class="fs-4 bi bi-trash"></i></a>
+                                    <a onclick="return confirm('Tem a certeza que quer apagar?');" href="router.php?c=produto&a=delete&id=<?= $produto->id ?>" class="text-black" title="Apagar"><i class="fs-4 bi bi-trash"></i></a>
                                 </td>
                             </tr>
                         <?php } ?>
